reduce memory footprint
Army size is now same as master calc if master calc remains unoptimized

+++ battleLogic.h
@@ -44,7 +44,7 @@ public:
 	int frontDamageTaken;
 	int healing;
 	int cumAoeDamageTaken;
-	float berserkProcs;
+	int berserkProcs;
 
 	// skill data
 	int damage;
